Rename RentalInfo class to RentalInformationService
