+++ src/main/java/com/mithwick93/refactoring/java/Main.java
@@ -4,6 +4,7 @@ import com.mithwick93.refactoring.java.entity.Customer;
 import com.mithwick93.refactoring.java.entity.MovieRental;
 import com.mithwick93.refactoring.java.repositroy.MovieRepository;
 import com.mithwick93.refactoring.java.repositroy.impl.MovieRepositoryImpl;
+import com.mithwick93.refactoring.java.service.RentalInformationService;
 
 import java.util.Arrays;
 
@@ -21,7 +22,7 @@ public final class Main {
                 """;
 
         MovieRepository movieRepository = new MovieRepositoryImpl();
-        String result = new RentalInfo(movieRepository)
+        String result = new RentalInformationService(movieRepository)
                 .statement(
                         new Customer(
                                 "C. U. Stomer",
